refactor: util 분리
